Review update, DragPointsItemInspector
child class of ItemInspector, replace DragPointsEditor
