--- VisualPinball.Unity/VisualPinball.Unity.Editor/Editors.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f438698183143fe42ae0a2120b4245d2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
