+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Inspectors/DragPointsItemInspector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e106fbedb26d4124e988bee28c4d2e6c
+guid: 0bd3f090848c3094c9ae70620f7d80a4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
